Security fixes: Sequence UI

+++ components/mediation-ui/org.wso2.carbon.sequences.ui/src/main/resources/web/sequences/delete_sequence.jsp
@@ -21,6 +21,11 @@
 <%@ page import="java.util.ResourceBundle" %>
 
 <%
+    //ignore methods other than post
+    if (!request.getMethod().equalsIgnoreCase("POST")) {
+        response.sendError(405);
+        return;
+    }
     String sequenceName = request.getParameter("sequenceName");
     String sequenceType = request.getParameter("type");
     boolean forceDelete = "true".equals(request.getParameter("force"));
